POJO finished + Schema added;

+++ src/main/java/com/spribe/project/models/request/delete/PlayerDeleteRequest.java
@@ -1,17 +1,15 @@
 package com.spribe.project.models.request.delete;
 
-import com.spribe.project.models.request.get.PlayerGetByPlayerIdRequest;
-
 import java.util.Objects;
 
 public class PlayerDeleteRequest {
-    private Integer playerId;
+    private Long playerId;
 
-    public Integer getPlayerId() {
+    public Long getPlayerId() {
         return playerId;
     }
 
-    public void setPlayerId(Integer playerId) {
+    public void setPlayerId(Long playerId) {
         this.playerId = playerId;
     }
